Revert "pure CSS arrow for external links, avoids problems with bad fonts"
This reverts commit f92680c739539eabb6d1e21e7a1fd15547546b0a.

It's not ideal. E.g. when hovering on a menu item with an external link,
it shows two black triangles, since the background has to be black for
the shadow drop to work.  There's gotta be a better way.  For now, I'm
reverting to this solution which works on most devices.